refactor: ban TYPE_CHECKING and break all circular imports
Eliminate all 22 `if TYPE_CHECKING:` usages across the codebase.

17 files had no actual circular dependency — imports moved to top level.
5 genuine cycles broken by restructuring:

- networking ↔ sync: use direct PeerId import path instead of package re-export
- block/types.py ↔ block/block.py: move BlockLookup class into block.py
- block/block.py ↔ state: narrow verify_signatures param from State to
  Validators (moved from state/types.py to validator.py where it belongs)
- xmss/containers.py ↔ interface.py: delete Signature.verify() convenience
  wrapper, callers use scheme.verify() directly
- stream_adapter.py ↔ connection.py: deferred import of QuicStreamAdapter
  after QuicStream is defined, so stream_adapter.py can import normally

Zero noqa/ignore suppressions. Zero Protocols. All types fully resolved.
