+++ CLAUDE.md
@@ -58,7 +58,7 @@ uvx tox                  # Everything (checks + tests + docs)
 
 ### Import Style
 
-**All imports must be at the top of the file.** Never place imports inside functions, methods, or conditional blocks. This applies to both source code and tests. If a circular dependency exists, restructure the code to break the cycle rather than using a lazy import.
+**All imports must be at the top of the file.** Never place imports inside functions, methods, or conditional blocks. This applies to both source code and tests. The **only** exception is genuine circular dependencies — in that case, import inside the function that needs the type (see the `TYPE_CHECKING` rule below).
 
 Bad:
 ```python
@@ -103,6 +103,40 @@ from cryptography.hazmat.primitives.asymmetric import x25519
 ```
 
 This keeps code readable and avoids mental overhead of tracking renamed imports.
+
+**CRITICAL - Never use `TYPE_CHECKING`.** The `if TYPE_CHECKING:` pattern is banned entirely from this codebase. Do not import `TYPE_CHECKING` from `typing`, and do not place any imports behind `if TYPE_CHECKING:` guards. This pattern is fragile, hard to maintain, and causes subtle bugs — especially with Pydantic models.
+
+Instead:
+
+- **No circular dependency?** Just import normally at the top of the file. Most guarded imports have no actual cycle.
+- **Genuine circular dependency?** Import inside the function that needs it. This is the **only** exception to the top-level import rule. Keep the local import as close as possible to where it's used.
+- **Forward references needed?** Use quoted strings (`"ClassName"`) in annotations, or add `from __future__ import annotations` (but NOT in Pydantic model files).
+
+Bad:
+```python
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from lean_spec.subspecs.forkchoice import Store
+
+def process(store: "Store") -> None: ...
+```
+
+Good:
+```python
+from lean_spec.subspecs.forkchoice import Store
+
+def process(store: Store) -> None: ...
+```
+
+Good (local import for genuine circular dependency):
+```python
+def verify_signatures(self, state: "State") -> bool:
+    from ..state import State
+
+    # use State here
+    ...
+```
 
 ### Type Annotations
