+++ packages/testing/src/consensus_testing/test_types/block_spec.py
@@ -2,8 +2,6 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
 from lean_spec.subspecs.containers.attestation import (
     AggregatedAttestation,
     Attestation,
@@ -21,20 +19,16 @@ from lean_spec.subspecs.containers.block.types import (
     AttestationSignatures,
 )
 from lean_spec.subspecs.containers.slot import Slot
+from lean_spec.subspecs.containers.state.state import State
 from lean_spec.subspecs.containers.validator import ValidatorIndex
+from lean_spec.subspecs.forkchoice.store import Store
 from lean_spec.subspecs.ssz.hash import hash_tree_root
 from lean_spec.subspecs.xmss.aggregation import AggregatedSignatureProof
+from lean_spec.subspecs.xmss.containers import Signature
 from lean_spec.types import Bytes32, CamelModel
 
-from ..keys import LEAN_ENV_TO_SCHEMES, create_dummy_signature
+from ..keys import LEAN_ENV_TO_SCHEMES, XmssKeyManager, create_dummy_signature
 from .aggregated_attestation_spec import AggregatedAttestationSpec
-
-if TYPE_CHECKING:
-    from lean_spec.subspecs.containers.state.state import State
-    from lean_spec.subspecs.forkchoice.store import Store
-    from lean_spec.subspecs.xmss.containers import Signature
-
-    from ..keys import XmssKeyManager
 
 
 class BlockSpec(CamelModel):
