+++ packages/testing/src/consensus_testing/test_types/gossip_aggregated_attestation_spec.py
@@ -2,23 +2,18 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
 from lean_spec.subspecs.containers.attestation import AttestationData
 from lean_spec.subspecs.containers.attestation.attestation import SignedAggregatedAttestation
+from lean_spec.subspecs.containers.block.block import Block
 from lean_spec.subspecs.containers.checkpoint import Checkpoint
 from lean_spec.subspecs.containers.slot import Slot
+from lean_spec.subspecs.containers.state.state import State
 from lean_spec.subspecs.containers.validator import ValidatorIndex, ValidatorIndices
 from lean_spec.subspecs.xmss.aggregation import AggregatedSignatureProof
 from lean_spec.types import ByteListMiB, Bytes32, CamelModel
 
+from ..keys import XmssKeyManager
 from .utils import resolve_checkpoint
-
-if TYPE_CHECKING:
-    from lean_spec.subspecs.containers.block.block import Block
-    from lean_spec.subspecs.containers.state.state import State
-
-    from ..keys import XmssKeyManager
 
 
 class GossipAggregatedAttestationSpec(CamelModel):
