+++ src/lean_spec/subspecs/storage/database.py
@@ -9,14 +9,13 @@ from __future__ import annotations
 
 from collections.abc import Generator
 from contextlib import contextmanager
-from typing import TYPE_CHECKING, Protocol
+from typing import Protocol
 
-if TYPE_CHECKING:
-    from lean_spec.subspecs.containers import Block, Checkpoint, State
-    from lean_spec.subspecs.containers.attestation import AttestationData
-    from lean_spec.subspecs.containers.slot import Slot
-    from lean_spec.subspecs.containers.validator import ValidatorIndex
-    from lean_spec.types import Bytes32, Uint64
+from lean_spec.subspecs.containers import Block, Checkpoint, State
+from lean_spec.subspecs.containers.attestation import AttestationData
+from lean_spec.subspecs.containers.slot import Slot
+from lean_spec.subspecs.containers.validator import ValidatorIndex
+from lean_spec.types import Bytes32, Uint64
 
 
 class Database(Protocol):
