+++ src/lean_spec/subspecs/sync/peer_manager.py
@@ -12,9 +12,9 @@ from dataclasses import dataclass, field
 from typing import Final
 
 from lean_spec.subspecs.containers.slot import Slot
-from lean_spec.subspecs.networking import PeerId
 from lean_spec.subspecs.networking.peer import PeerInfo
 from lean_spec.subspecs.networking.reqresp.message import Status
+from lean_spec.subspecs.networking.transport.peer_id import PeerId
 
 from .config import MAX_CONCURRENT_REQUESTS
